show error response for server exceptions as well as client exceptions
